+++ CHANGELOG.md
@@ -37,6 +37,7 @@ The format is based on [Keep a Changelog]
     - Add 'What gender does your school's payroll system associate with you?'
     - Add 'Enter bank account details'
     - Add 'Email address'
+  - Add 'Application complete'
 - Automatically update admin claim tasks with DQT API
 
 ## [Release 089] - 2021-03-05